refactor: move cache logic to only encode driver
